requirements.txt
frame-msg
opencv-python
numpy
Pillow
bleak 